+++ crates/ryo3-bytes/src/ryo3_bytes.rs
@@ -1,6 +1,7 @@
 //! Support for Python buffer protocol
 
 use std::fmt::Write;
+use std::ops::Range;
 use std::os::raw::c_int;
 use std::ptr::NonNull;
 
@@ -12,7 +13,9 @@ use pyo3::pybacked::PyBackedBytes;
 use pyo3::types::{PyDict, PySlice, PyString, PyTuple};
 use pyo3::{IntoPyObjectExt, ffi};
 
-use crate::python_bytes_methods::PythonBytesMethods;
+use crate::ReadableBuffer;
+use crate::python_bytes_methods::{PyHexSep, PythonBytesMethods, PythonBytesStrip};
+use crate::replace::{ReplaceBytes, replace_bytes};
 
 /// A wrapper around a [`bytes::Bytes`][].
 ///
@@ -68,6 +71,11 @@ impl PyBytes {
     /// Access the underlying buffer as a byte slice
     pub fn as_slice(&self) -> &[u8] {
         self.as_ref()
+    }
+
+    /// Return the range of the buffer
+    pub fn range(&self) -> Range<usize> {
+        0..self.0.len()
     }
 
     /// Slice the underlying buffer using a Python slice object
@@ -191,14 +199,14 @@ impl PyBytes {
         new_buffer.into()
     }
 
-    fn __contains__(&self, item: Self) -> bool {
+    fn __contains__(&self, item: ReadableBuffer) -> bool {
         self.0
-            .windows(item.0.len())
-            .any(|window| window == item.as_slice())
+            .windows(item.len())
+            .any(|window| window == item.as_ref())
     }
 
-    fn __eq__(&self, other: Self) -> bool {
-        self.0.as_ref() == other.0.as_ref()
+    fn __eq__(&self, other: ReadableBuffer) -> bool {
+        self.0.as_ref() == other.as_ref()
     }
 
     fn __getitem__<'py>(
@@ -272,9 +280,9 @@ impl PyBytes {
     /// If the binary data starts with the prefix string, return bytes[len(prefix):]. Otherwise,
     /// return a copy of the original binary data:
     #[pyo3(signature = (prefix, /))]
-    fn removeprefix(&self, prefix: Self) -> Self {
+    fn removeprefix(&self, prefix: ReadableBuffer) -> Self {
         if self.0.starts_with(prefix.as_ref()) {
-            self.0.slice(prefix.0.len()..).into()
+            self.0.slice(prefix.len()..).into()
         } else {
             self.0.clone().into()
         }
@@ -283,11 +291,24 @@ impl PyBytes {
     /// If the binary data ends with the suffix string and that suffix is not empty, return
     /// `bytes[:-len(suffix)]`. Otherwise, return the original binary data.
     #[pyo3(signature = (suffix, /))]
-    fn removesuffix(&self, suffix: Self) -> Self {
+    fn removesuffix(&self, suffix: ReadableBuffer) -> Self {
         if self.0.ends_with(suffix.as_ref()) {
-            self.0.slice(0..self.0.len() - suffix.0.len()).into()
+            self.0.slice(0..self.0.len() - suffix.len()).into()
         } else {
             self.0.clone().into()
+        }
+    }
+
+    #[pyo3(signature = (old, new, count = -1, /))]
+    fn replace(
+        slf: PyRef<'_, Self>,
+        old: ReadableBuffer,
+        new: ReadableBuffer,
+        count: isize,
+    ) -> PyResult<Py<Self>> {
+        match replace_bytes(slf.as_slice(), old.as_slice(), new.as_slice(), count) {
+            ReplaceBytes::Unchanged => Ok(slf.into()),
+            ReplaceBytes::Replaced(bytes) => Py::new(slf.py(), Self::from(bytes)),
         }
     }
 
@@ -383,6 +404,10 @@ impl PyBytes {
         self.py_hash()
     }
 
+    fn __iter__(&self) -> PyBytesIterator {
+        self.into()
+    }
+
     /// Decode the bytes using the codec registered for encoding.
     ///
     ///   encoding
@@ -426,9 +451,9 @@ impl PyBytes {
     /// 'b9:01ef'
     /// >>> value.hex(':', -2)
     /// 'b901:ef'
-    #[pyo3(signature = (sep = None, bytes_per_sep = None))]
-    fn hex(&self, sep: Option<&str>, bytes_per_sep: Option<usize>) -> PyResult<String> {
-        self.py_hex(sep, bytes_per_sep)
+    #[pyo3(signature = (sep = None, *, bytes_per_sep = 1))]
+    fn hex(&self, sep: Option<PyHexSep>, bytes_per_sep: usize) -> String {
+        self.py_hex(sep.map(char::from), bytes_per_sep)
     }
 
     /// Create a bytes object from a string of hexadecimal numbers.
@@ -505,12 +530,12 @@ impl PyBytes {
     }
 
     #[pyo3(signature = (prefix, /))]
-    fn startswith(&self, prefix: Self) -> bool {
+    fn startswith(&self, prefix: ReadableBuffer) -> bool {
         self.as_slice().starts_with(prefix.as_ref())
     }
 
     #[pyo3(signature = (suffix, /))]
-    fn endswith(&self, suffix: Self) -> bool {
+    fn endswith(&self, suffix: ReadableBuffer) -> bool {
         self.as_slice().ends_with(suffix.as_ref())
     }
 
@@ -527,30 +552,36 @@ impl PyBytes {
         self.py_expandtabs(tabsize)
     }
 
-    #[pyo3(signature = (chars = None, /))]
-    fn strip(&self, chars: Option<Self>) -> Self {
-        if let Some(chars) = chars {
-            self.py_strip(Some(chars.as_ref()))
+    #[pyo3(signature = (chars = PythonBytesStrip::AsciiWhitespace, /), text_signature = "(chars=None, /)")]
+    fn strip(slf: PyRef<'_, Self>, chars: PythonBytesStrip) -> PyResult<Py<Self>> {
+        let bytes = &slf.0;
+        let range = chars.strip_range(slf.as_slice());
+        if range.start == 0 && range.end == bytes.len() {
+            Ok(slf.into())
         } else {
-            self.py_strip(None)
+            Py::new(slf.py(), Self::new(bytes.slice(range)))
         }
     }
 
-    #[pyo3(signature = (chars = None, /))]
-    fn lstrip(&self, chars: Option<Self>) -> Self {
-        if let Some(chars) = chars {
-            self.py_lstrip(Some(chars.as_ref()))
+    #[pyo3(signature = (chars = PythonBytesStrip::AsciiWhitespace, /), text_signature = "(chars=None, /)")]
+    fn lstrip(slf: PyRef<'_, Self>, chars: PythonBytesStrip) -> PyResult<Py<Self>> {
+        let bytes = &slf.0;
+        let ix = chars.lstrip_range(slf.as_slice());
+        if ix == 0 {
+            Ok(slf.into())
         } else {
-            self.py_lstrip(None)
+            Py::new(slf.py(), Self::new(bytes.slice(ix..)))
         }
     }
 
-    #[pyo3(signature = (chars = None, /))]
-    fn rstrip(&self, chars: Option<Self>) -> Self {
-        if let Some(chars) = chars {
-            self.py_rstrip(Some(chars.as_ref()))
+    #[pyo3(signature = (chars = PythonBytesStrip::AsciiWhitespace, /), text_signature = "(chars=None, /)")]
+    fn rstrip(slf: PyRef<'_, Self>, chars: PythonBytesStrip) -> PyResult<Py<Self>> {
+        let bytes = &slf.0;
+        let ix = chars.rstrip_range(slf.as_slice());
+        if ix == bytes.len() {
+            Ok(slf.into())
         } else {
-            self.py_rstrip(None)
+            Py::new(slf.py(), Self::new(bytes.slice(0..ix)))
         }
     }
     // </python-bytes-methods>
@@ -667,4 +698,37 @@ enum BytesGetItemKey<'py> {
     Int(isize),
     /// A python slice
     Slice(Bound<'py, PySlice>),
+}
+
+/// optimized bytes-iterator...
+
+#[pyclass(name = "BytesIterator", immutable_type)]
+#[cfg_attr(feature = "ry", pyo3(module = "ry.ryo3"))]
+pub(crate) struct PyBytesIterator(::bytes::buf::IntoIter<Bytes>);
+
+impl From<Bytes> for PyBytesIterator {
+    fn from(value: Bytes) -> Self {
+        Self(value.into_iter())
+    }
+}
+
+impl From<&PyBytes> for PyBytesIterator {
+    fn from(value: &PyBytes) -> Self {
+        value.0.clone().into()
+    }
+}
+
+#[pymethods]
+impl PyBytesIterator {
+    fn __iter__(slf: PyRef<'_, Self>) -> PyRef<'_, Self> {
+        slf
+    }
+
+    fn __next__(&mut self) -> Option<u8> {
+        self.0.next()
+    }
+
+    fn __len__(&self) -> usize {
+        self.0.len()
+    }
 }